Use separate classes to hold token and element types
This makes it easier to tell token and element types apart.
Unfortunately Grammar-Kit cannot generate token and element types into
separate classes by itself.
