+++ src/gen/java/com/atslangplugin/psi/impl/ATSDummyImpl.java
@@ -7,7 +7,7 @@ import com.intellij.lang.ASTNode;
 import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiElementVisitor;
 import com.intellij.psi.util.PsiTreeUtil;
-import static com.atslangplugin.psi.ATSTokenTypes.*;
+import static com.atslangplugin.psi.ATSElementTypes.*;
 import com.intellij.extapi.psi.ASTWrapperPsiElement;
 import com.atslangplugin.psi.*;
